requirements.txt
streamlit>=1.31.0
pymupdf>=1.23.0
pytesseract>=0.3.0
huggingface-hub>=0.20.0
openpyxl>=3.1.0
pandas>=2.0.0
whoosh>=2.7.4
Pillow>=10.0.0
python-dotenv>=1.0.0
